backend/pyproject.toml
[project]
name = "ai-agent-backend"
version = "0.1.0"
description = "Backend for an AI agent using FastAPI and OpenAI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "openai>=1.3",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "SQLAlchemy[asyncio]>=2.0",
    "alembic>=1.11",
    "httpx>=0.25",
    "apscheduler>=3.10",
]

[tool.pdm]
allow_prereleases = true